requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-cors-headers==3.8.0
redis==4.0.3
celery==5.2.2
psycopg2-binary==2.9.1
python-memcached==1.59